Upload FFLStats Source
I designed this program for tracking and editing the statistics of
football players for my fantasy football league. Initially I had the
stats saved in a .txt file and editing was a hassle with notepad. This
program takes the .txt file and places the players and their stats in a
struct named footBallPlayers. Once the file has been read and all values
are stored the user is prompted with a menu for instructions. With this
menu the user is able to view all players' stats, view a particular
player's stats, view players by position, add to the player's stats,
edit the player's stats and position, or exit. The user isn't limited to
one choice, once a process is finished the user is prompted with the
menu again. Upon deciding to exit the user is prompted to save the
changes to the .txt file or exit without saving.

The menu function is recursive which saved lines and time. Once the user
makes a selection of what they wish to do, the selected function
executes then calls the menu function again.